遷移アクションをFragmentで管理するようにしたのでRecyclerViewAdapterをFragmentに引っ越し
